get proper height for legend
