+++ ui/components/src/Table/index.ts
@@ -1,1 +1,2 @@
 export * from './Table';
+export { getTableCellLayout } from './table-model';
