+++ ui/panels-plugin/src/plugins/time-series-chart/utils/layout-utils.test.ts
@@ -11,6 +11,9 @@ describe('getTimeSeriesLayout', () => {
         },
         showYAxis: false,
         spec: {},
+
+        // @ts-ignore
+        theme: {},
       })
     ).toMatchSnapshot();
   });
